Speed up SimpleStorage E2E significantly
This takes advantage of docker-compose not building if an image with the
same name exists locally. Since Travis is building on Linux, the
binaries that get produced by a stack build are compatible and can just
be copied wholesale into the container. This significantly speeds up
E2E in Travis as we don't have to rebuild the whole app inside Docker
and can take advantage of Travis' .stack-work cache as well. In the
travis script, we run docker build with the appropriate tag to create
the image locally, and all docker-compose.yamls which use these services
can take advantage of that prebaked image.