A more thorough CI pipeline (#116)
* break up .travis.yml into build stages

* add a stage for testing hs-tendermint-client against abci-cli kvstore and supporting docker stuff

* fix compose for kvstore test, use stages/job names correctly for parallelism

* potential fix

* add requisite flags and versions to tendermint to make kvstore test pass

* simplify stage layout, create simplestorage e2e

* fix travis.yml syntax

* use docker volumes instead of /tmp directory, use consistent naming in composes, move disabled nix test into its own job for when its ready to be used

* move Dockerfile to project-wide level, fix dockerignore to not copy stack stuff, have simple-storage E2E compose build using project-wide Dockerfile

* Speed up SimpleStorage E2E significantly

This takes advantage of docker-compose not building if an image with the
same name exists locally. Since Travis is building on Linux, the
binaries that get produced by a stack build are compatible and can just
be copied wholesale into the container. This significantly speeds up
E2E in Travis as we don't have to rebuild the whole app inside Docker
and can take advantage of Travis' .stack-work cache as well. In the
travis script, we run docker build with the appropriate tag to create
the image locally, and all docker-compose.yamls which use these services
can take advantage of that prebaked image.

* expose the correct port in simple-storage docker-compose so it can actually run

* change both queries to use client

* merged client-store-fix

* fix unused imports

* stylish